applications/travel_planning_agent/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<4
sqlalchemy==2.0.43
python-multipart==0.0.20
httpx==0.28.1
genxai-framework==0.1.5
openai==2.20.0
python-dotenv==1.1.1
